car_rental_api/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
email-validator==1.3.1
python-multipart==0.0.9
